+++ qulab/dht/network.py
@@ -12,8 +12,8 @@ from qulab._config import config
 from qulab.dht.crawling import NodeSpiderCrawl, ValueSpiderCrawl
 from qulab.dht.node import Node
 from qulab.dht.protocol import KademliaProtocol
-from qulab.dht.storage import ForgetfulStorage
 from qulab.dht.utils import digest
+from qulab.storage import ForgetfulStorage
 
 cfg = config.get('dht', dict())
 DEFALT_PORT = cfg.get('default_port', 8987)
