+++ qulab/storage/__init__.py
@@ -1,0 +1,1 @@
+from qulab.storage.memstorage import ForgetfulStorage
